Remove old roles data
