+++ src/_data/roles.js
@@ -1,191 +1,54 @@
-module.exports = [
-    {
-        group: "Artificial Intelligence engineers",
-        roles: [
-            {
-                title: "Applied Artificial Intelligence engineer",
-                roleSpec: [
-                    "Collaborate with cross-functional teams including researchers, product managers and senior stakeholders to quickly understand customer needs and translate them into pragmatic technical solutions leveraging AI.",
-                    "Rapidly prototype AI-powered features and iterate based on user feedback.",
-                    "Optimise AI models and infrastructure for performance, scalability and reliability.",
-                    "Establish frameworks, tools and processes to responsibly develop, deploy and monitor AI systems.",
-                    "Evangelise best practices in applied AI across government."
-                ],
-                personSpec: [
-                    "Proficiency in Python and experience with the full data science and machine learning stack (Pandas, Numpy, PyTorch, Langchain, Pydantic).",
-                    "Strong software engineering skills with a focus on building robust, production ready systems and understanding of deploying and monitoring such systems in a cloud environment.",
-                    "Experience with natural language processing, deep learning and generative AI including prompt engineering.",
-                    "Passion for building AI that augments human capabilities and improves citizen’s lives."
-                ]  
-            }
-        ]
-    },
-    {
-        group: "Data",
-        roles: [
-            {
-                title: "Data engineer",
-                roleSpec: [
-                    "Work closely with AI engineers to understand infrastructure needs and bottlenecks.",
-                    "Implement and maintain cloud infrastructure solutions in accordance with good practice and departmental policy.",
-                    "Monitor and understand resource utilisation, make recommendations, manage optimisation and scaling to ensure resources are used efficiently and cost-effectively.",
-                    "Design infrastructure solutions using Kubernetes, Terraform, Serverless functions like lambdas, SQS, S3 and other services.",
-                    "Collaborate across the civil service working with existing technical solutions and understand how to best build into existing architectures."
-                ],
-                personSpec: [
-                    "Deep experience in infrastructure engineering, DevOps or large scale data platforms.",
-                    "Expert in Infrastructure-as-code (Terraform, Cloudformation) and experience with moving legacy infrastructure into state.",
-                    "Strong AWS experience including creating and maintaining VPC’s, security groups, network ACL’s, subnets, & nat gateways with a focus on creating private infrastructure with controlled access.",
-                    "Knowledge of containerisation, scripting languages (Bash) including Python and the ability to setup and architect CI/CD pipelines."
+require('dotenv').config();
+const contentful = require('contentful');
+
+const client = contentful.createClient({
+    space: process.env.CONTENTFUL_SPACE,
+    accessToken: process.env.CONTENTFUL_ACCESS_TOKEN
+});
   
-                ] 
+module.exports = () => {
+
+    return client.getEntries({
+        content_type: 'roles'
+    }).then((response) => {
+
+        let roles = response.items.map((role) => {
+            return {
+                title: role.fields.title,
+                group: role.fields.group.fields.name,
+                roleSpec: role.fields.roleResponsibilities ? Object.values(role.fields.roleResponsibilities).map(responsibility => responsibility.fields.text) : [],
+                personSpec: Object.values(role.fields.personSpecification).map(specification => specification.fields.text)
+            };
+        });
+
+        // sort by group, then by title
+        roles.sort((a, b) => {
+            if (a.group === b.group) {
+                return a.title < b.title ? -1 : 1;
             }
-        ]
-    },
-    {
-        group: "Evaluation",
-        roles: [
-            {
-                title: "AI Impact and Evaluation Lead",
-                roleSpec: [
-                    "Design and implement a new AI platform architecture from the ground up focused on scalability, modularity and performance.",
-                    "Develop infrastructure to support running large scale AI training and inference jobs and distributed model training.",
-                    "Create environments optimised for GPU/TPU workloads and AI-specific hardware.",
-                    "Work closely with engineers and researchers to improve infrastructure for scalable AI applications.",
-                    "Evaluate and experiment with new tools and technologies to improve speed and aid efficiency of the AI stack."
-                ],
-                personSpec: [
-                    "Experience in creating and managing HPC clusters across GPU/TPU and with job schedulers like Slurm and Fsx volumes.",
-                    "Knowledge of cloud computing platforms, specifically AWS and experience with infrastructure as code tools like Terraform.",
-                    "Experience in machine learning for production using Python and related tools (PyTorch, webdataset, torchdata).",
-                    "Understanding of large scale model inference, quantization and distributed computing.",
-                    "Mindset to build reusable, modular infrastructure and systems focused on scaling."
-                ]
+            return a.group < b.group ? -1 : 1;
+        });
+
+        // split into groups
+        const groups = [];
+        roles.forEach((role) => {
+            if (groups.length && groups[groups.length - 1].group === role.group) {
+                groups[groups.length - 1].roles.push(role);
+            } else {
+                groups.push({
+                    group: role.group,
+                    roles: [role]
+                });
             }
-        ]
-    },
-    {
-        group: "Product and delivery",
-        roles: [
-            {
-                title: "Delivery Manager",
-                roleSpec: [
-                    "Designing the appropriate agile practices for the team based on the needs of the team and the work.",
-                    "Coaching and mentoring others in approaches to agile delivery.",
-                    "Managing dependencies between projects and escalating issues you or the team identify.",
-                    "Managing risks, commercials and budgets to ensure resources are used efficiently.",
-                    "Working with the team to ensure products are scoped appropriately to deliver a MVP that meets stakeholder’s and user’s needs while maintaining focus on prototyping.",
-                    "Identifying, sourcing and managing the tools needed for the team.",
-                    "Line managing people to enable to deliver their best work.",
-                    "Supporting with the creation of business cases as required for the team’s work.",
-                    "Proactively finding opportunities to make the team more inclusive and supporting the team’s well being.",
-                    "Embedding a user centred design ways of working to the team."
-                ],
-                personSpec: [
-                    "Agile practices and maintaining momentum- You can coach and lead teams in Agile and Lean practices. You can create or tailor new ways of working, and constantly innovate. You can actively address the most complicated risks, issues and dependencies including where ownership exists outside the team or no clear ownership exists. You can identify innovative ways to unblock issues.",
-                    "Life cycle perspective- You are experienced across multiple parts of the product life cycle. You can recognise when it is right to move forward and when it is right to stop. You can recognise the appropriate deliverables and the right people to meet them. You can plan and engage with the appropriate stakeholders at a particular stage in the project.",
-                    "Making a process work- You can identify and improve organisational processes of increasing complexity and those processes that are unnecessarily complicated.",
-                    "Team dynamics and collaboration- You work to improve the diversity and wellbeing of the team. You can identify problems or issues in the team dynamic and rectify them. You can engage in varying types of feedback, choosing the right type at the appropriate time. You can accelerate the team development cycle.",
-                    "Understanding and documenting needs- You can quickly identify the needs of a product or service and support the Head of Delivery to identify where to start. You consider what’s best for the user, the organisation and the team.",
-                    "Commercial and financial management- You can identify what roles the team needs and take responsibility for ensuring they are in place relationships with contracted suppliers and HR. You can get good value out of contracts and suppliers. You can write or input into budgets and business cases and can communicate business-value propositions.",
-                    "Communicating between the technical and non-technical- You can mediate between people and build relationships, communicating with stakeholders at all levels. You can manage stakeholder expectations and moderate discussions about high risk and complexity, even within constrained timescales. You can speak on behalf of and represent the community to large audiences inside and outside of government."
-                ]
-            },
-            {
-                title: "Service Owner",
-                roleSpec: [
-                    "Operate at scale and provide the connection between multidisciplinary business areas and stakeholders.",
-                    "Ensure the necessary business processes are followed.",
-                    "Participate in the governance of the service, including acting as a point of escalation for the delivery teams.",
-                    "Own the budget and allocate funding to areas of the service based on your decisions about priorities.",
-                    "Communicate the benefits and performance of your service.",
-                    "Be responsible for the successful operation and continuous improvement of the service."
-                ],
-                personSpec: [
-                    "Coaching the team on agile practices with a clear focus on MVP.",
-                    "Experience of leading a team to delivery though stages of the product life cycle.",
-                    "Leading a team to operationalise a products, handling risk and being a point of escalation.",
-                    "Risk and problem management, evidence of seeing how the product fits into the bigger picture and making effective decisions.",
-                    "Owning the strategic roll out of a product of service.",
-                    "Apply strategic thinking to provide the best service for the end user."
-                ]
-            }
-        ]
-    },
-    {
-        group: "Software development",
-        roles: [
-            {
-                title: "Frontend developer and accessibility expert",
-                roleSpec: [
-                    "Leading and delivering frontend development in the team, setting standards for frontend development and building capability in the team.",
-                    "Suggesting accessible design solutions, and championing accessibility in the team.",
-                    "Leading projects, and managing the work of others in the team.",
-                    "Working with product managers and stakeholders to understand user needs and translate these into technical solutions.",
-                    "Working flexibly to support changing priorities.",
-                    "Supporting the i.AI Head of Delivery and Deputy Director in maintaining the wellbeing of the team, and ensuring it is an inclusive place to work."
-                ],
-                personSpec: [
-                    "Experienced frontend developer (JavaScript, HTML, CSS).",
-                    "Writing clean, maintainable code and tests.",
-                    "Development of a UI design system library.",
-                    "Writing accessible code following a progressive enhancement approach.",
-                    "Understanding of how to design accessible solutions.",
-                    "Knowledge of tools and methods to use when accessibility testing.",
-                    "Ability to translate user needs into requirements.",
-                    "Experience mentoring and coaching junior developers.",
-                    "Proactive in learning new skills and upskilling others."
-                ]
-            }
-        ]
-    },
-    {
-        group: "User-centred design",
-        roles: [
-            {
-                title: "User researcher",
-                roleSpec: [
-                    "Understanding user needs and working with the team and stakeholders to use those insights to make decisions about the products we build.",
-                    "Working with the team to identify assumptions, designing and delivering research to prove/disprove them.",
-                    "Designing research plans on how we can test and learn from users so we gain valuable insight on the problem space.",
-                    "Carrying out research, this could be user interviews, desk research, analysis of survey responses and more.",
-                    "Coaching the team and others in User Centred Design.",
-                    "Recording insights and methods from research conducted to evidence ways of work and how we meet the service standard."
-                ],
-                personSpec: [
-                    "Research. You can demonstrate experience in a wide range of user research methods, and can help teams to adopt them. You can plan user research for services with challenging user needs and complex user journeys. You can advise colleagues on the choice and application of research methods to assure best practice.",
-                    "Inclusive research. You can help teams understand the diversity of users of government services. You can effectively include all kinds of users in appropriate research activities. You can advocate inclusive practices and help teams design and deliver accessible services that work for all users.",
-                    "Analysis and synthesis. You can understand and help teams to apply a range of methods to analyse research data and synthesise findings. You can effectively engage sceptical colleagues in analysis and synthesis. You can advise on the choice and application of techniques, and can critique colleagues’ findings to assure best practice.",
-                    "Strategic insight. You can understand what problem the team is trying to solve and can align user research activities to inform decision making and action.",
-                    "Communicating between the technical and non-technical. You can mediate between people and build relationships, communicating with stakeholders at all levels, listen to the needs of technical and business stakeholders, and interpret them effectively to manage stakeholder expectations.",
-                    "Collaboration. You can work collaboratively in a group, actively networking with others. You can adapt feedback to ensure it’s effective and usable. You use your initiative to identify problems or issues in the team dynamic and rectify them."
-                ]
-            },
-            {
-                title: "Service designer",
-                personSpec: [
-                    "User-centred design- You show a deep understanding of end-to-end journeys and how content is affected within these journeys. You will be accountable for the production of high quality, user-focused content. You ensure that content and design is regularly reviewed and evaluated, contributing to continuous improvements and iteration of products.",
-                    "Prototyping- You can build and create a variety of different prototypes, from paper sketches to coded prototypes. You can choose the most appropriate option to visualise content in context for different audiences. You can effectively combine prototypes with other content and design outputs to test with users and show to stakeholders. You can establish content / design patterns, and test and iterate them.",
-                    "Communicating between the technical and non-technical- You can mediate between people and build relationships, communicating with stakeholders at all levels. listen to the needs of technical and business stakeholders, and interpret them effectively manage stakeholder expectations.",
-                    "Content design- You can design content to meet user needs and make complex language and processes easy to understand. present the right information in the right format for what users need. You write in plain language in a way that users understand, making information accessible to all. You can use data, research and evidence to review and evaluate content to make improvements.",
-                    "Collaboration- You can work collaboratively in a group, actively networking with others. You can adapt feedback to ensure it’s effective and useable. You use your initiative to identify problems or issues in the team dynamic and rectify them."
-                ]
-            },
-            {
-                title: "Interaction designer",
-                roleSpec: [
-                    "Work with service managers and programme directors to develop design concepts.",
-                    "Set direction and embed good practice within the team.",
-                    "Create patterns for how users interaction with Artificial Intelligence.",
-                    "Contribute to design communities.",
-                    "Make important decisions based on research and understand how this research impacts others."
-                ],
-                personSpec: [
-                    "Leader in communicating between the technical and non-technical and designing solutions to meet business and user needs.",
-                    "Work collaboratively in a group, actively networking with others.",
-                    "Understanding user needs, including those with diverse needs and designing solutions that meet those needs."
-                ]
-            }
-        ]
-    }
-]
+        });
+
+        return groups;
+
+    }).catch((err) => {
+        console.error(err);
+    });
+
+};
+
+return [];
+
